% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as_simstudy}
\alias{as_simstudy}
\title{Translate Spec + Params to simstudy Arguments}
\usage{
as_simstudy(spec, params, type = "rand_effects")
}
\arguments{
\item{spec}{An \code{lmm_spec} object.}

\item{params}{An \code{lmm_params} object (fully set).}

\item{type}{One of \code{"rand_effects"} (args for \code{addCorData()}) or
\code{"outcome"} (args for \code{defDataAdd()}).}
}
\value{
A list of arguments for the corresponding simstudy function.
}
\description{
Translate Spec + Params to simstudy Arguments
}
