% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{lmm_params}
\alias{lmm_params}
\title{Create an Empty LMM Params Object}
\usage{
lmm_params(spec)
}
\arguments{
\item{spec}{An \code{lmm_spec} object.}
}
\value{
An object of class \code{lmm_params}.
}
\description{
Inspects an \code{lmm_spec} and returns a params object with all required slots
set to NULL. Prints a summary of what needs to be filled via \code{patch()}.
}
