% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{set_fixed}
\alias{set_fixed}
\title{Set Fixed Effects}
\usage{
set_fixed(spec, formula)
}
\arguments{
\item{spec}{An \code{lmm_spec} object.}

\item{formula}{A formula, e.g. \code{y ~ 1 + x1}. The response name overrides
the one set in \code{lmm_spec()}.}
}
\value{
An updated \code{lmm_spec}.
}
\description{
Set Fixed Effects
}
